Исправлены типы данных полей связанных записей в таблицах

+++ database/migrations/2024_12_05_194024_create_movies_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('movies', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('state_id')->nullable();
+            $table->unsignedBigInteger('state_id')->nullable();
             $table->string('poster_path')->nullable();
 
             $table->timestamps();
